pyproject.toml
[tool.poetry]
name = "pymidscene"
version = "0.1.0"
description = "Python SDK for Midscene.js - AI-powered UI automation using natural language"
authors = ["AIPythoner"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/AIPythoner/pymidscene"
repository = "https://github.com/AIPythoner/pymidscene"
documentation = "https://midscenejs.com"
keywords = ["midscene", "ai", "automation", "ui-testing", "playwright", "natural-language", "browser-automation", "vision-llm", "rpa"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Internet :: WWW/HTTP :: Browsers",
]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.0.0"
playwright = "^1.40.0"
pyyaml = "^6.0"
pillow = "^10.0.0"
httpx = "^0.25.0"
json-repair = "^0.7.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
typing-extensions = "^4.8.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
black = "^23.0.0"
mypy = "^1.5.0"
ruff = "^0.1.0"
types-pyyaml = "^6.0.0"

[tool.poetry.scripts]
pymidscene = "pymidscene.cli.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["json_repair.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --cov=pymidscene --cov-report=term-missing"
testpaths = ["tests"]
pythonpath = ["."]

[tool.coverage.run]
source = ["pymidscene"]
omit = ["tests/*", "examples/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
